Componentize buttons
